# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PCPartPicker API
  version: 1.0.0
  description: |
    REST API for the PCPartPicker prototype. The client must never access the Postgres/Supabase
    database directly; all reads and writes go through this API. Each resource exposes standard
    CRUD operations (create, read, update, delete) for the project’s tables.
servers:
  - url: http://localhost:3000
    description: Local dev server
paths:
...
  /vehicles:
    get:
      summary: List vehicles
      tags: [Vehicles]
      parameters:
        - in: query
          name: make
          schema:
            type: string
          description: Filter by make (e.g., Toyota)
        - in: query
          name: model
          schema:
            type: string
          description: Filter by model
        - in: query
          name: year
          schema:
            type: integer
          description: Filter by year
      responses:
        '200':
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vehicle'
    post:
      summary: Create a vehicle
      tags: [Vehicles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleCreate'
      responses:
        '201':
          description: Created vehicle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
  /vehicles/{vehicleId}:
    get:
      summary: Get vehicle by ID
      tags: [Vehicles]
      parameters:
        - $ref: '#/components/parameters/VehicleIdParam'
      responses:
        '200':
          description: Vehicle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '404':
          description: Not found
    patch:
      summary: Update vehicle
      tags: [Vehicles]
      parameters:
        - $ref: '#/components/parameters/VehicleIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleUpdate'
      responses:
        '200':
          description: Updated vehicle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
    delete:
      summary: Delete vehicle
      tags: [Vehicles]
      parameters:
        - $ref: '#/components/parameters/VehicleIdParam'
      responses:
        '204':
          description: Deleted

  # -------------------------
  # LISTINGS
  # -------------------------
...